Alta de cliente

+++ src/main/java/com/kennedy/erp/repositories/TipoClienteRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface TipoClienteRepository extends JpaRepository<TipoCliente, Long> {
+public interface TipoClienteRepository extends JpaRepository<TipoCliente, Integer> {
 }
